% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHIPRNAclass.R
\docType{class}
\name{chip}
\alias{chip}
\alias{chip-class}
\title{A ChIP-seq bed file of 4 columns of chr, start, end and clus}
\description{
Store a ChIP-seq peak file
}
\section{Slots}{

\describe{
\item{\code{bed}}{A \code{tbl} object}
}}

