% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHIPRNAclass.R
\docType{class}
\name{chiploj}
\alias{chiploj}
\alias{chiploj-class}
\title{Store ChIP-seq with binnarized intersections}
\description{
Store ChIP-seq with binnarized intersections
}
\section{Slots}{

\describe{
\item{\code{binmat}}{A \code{tbl} object}
}}

