% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHIPRNAconst.R
\name{chiplojConst}
\alias{chiplojConst}
\title{Create a \code{chiploj} object from binnarized peak intersection file}
\usage{
chiplojConst(chipF, reverse = FALSE)
}
\arguments{
\item{chipF}{Bed file with binnarized intersections in additional columns}

\item{reverse}{Reverse order of binnarized categories, when TRUE.}
}
\value{
a \code{chiploj} object
}
\description{
Create a \code{chiploj} object from binnarized peak intersection file
}
