% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHIPRNAgeneric.R, R/CHIPRNAmethods.R
\docType{methods}
\name{closeGene2Peak}
\alias{closeGene2Peak}
\alias{closeGene2Peak,chip-method}
\title{find nearest gene to peak}
\usage{
closeGene2Peak(obj,
  genef = "~/athena/Gencode/mm10/annotation/transcript_tss.bed",
  genomef = "~/athena/Gencode/mm10/sequence/chrNameLength.genome")

\S4method{closeGene2Peak}{chip}(obj,
  genef = "~/athena/Gencode/mm10/annotation/transcript_tss.bed",
  genomef = "~/athena/Gencode/mm10/sequence/chrNameLength.genome")
}
\arguments{
\item{obj}{A \code{chip} object}

\item{genef}{A genef (gene or tss) bed formatted file}

\item{genomef}{A genome file named *.genome}
}
\value{
An \code{gene2peak} object
}
\description{
find nearest gene to peak
}
