% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHIPRNAgeneric.R, R/CHIPRNAmethods.R
\docType{methods}
\name{clusing}
\alias{clusing}
\alias{clusing,data.frame,character-method}
\title{clustering of expressions of genes that are closest to peaks of interest}
\usage{
clusing(dat, pdffout)

\S4method{clusing}{data.frame,character}(dat, pdffout)
}
\value{
pr_mb predicted cluster membership in integers
}
\description{
clustering of genes that are closest to peaks of interest
}
\section{Slots}{

\describe{
\item{\code{dat}}{A \code{data.frame} object}

\item{\code{pdffout}}{A pdf file output file.}
}}

