% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHIPRNAgeneric.R, R/CHIPRNAmethods.R
\docType{methods}
\name{deDupGene}
\alias{deDupGene}
\alias{deDupGene,gene2peak-method}
\title{deduplicate the same gene closest to multiple peaks}
\usage{
deDupGene(obj, distThresh)

\S4method{deDupGene}{gene2peak}(obj, distThresh)
}
\arguments{
\item{obj}{A \code{gene2peak} object}

\item{distThresh}{A distance threshold for assigning genes to peaks.}
}
\value{
An updated \code{gene2peak} object
}
\description{
deduplicate the same gene closest to multiple peaks
}
