% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHIPRNAclass.R
\docType{class}
\name{gene2peak}
\alias{gene2peak}
\alias{gene2peak-class}
\title{A ChIP-seq bed file (4 columns) with 2 additional columns of gid and dist}
\description{
The \code{gene2peak} object may be initialized before/after filtering of gid and/or dist.
}
\section{Slots}{

\describe{
\item{\code{bed}}{A \code{tbl} object}
}}

