% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHIPRNAclass.R
\docType{class}
\name{limma}
\alias{limma}
\alias{limma-class}
\title{Linking peak interval files to limma differential statistics}
\description{
Linking peak interval files to limma differential statistics
}
\section{Slots}{

\describe{
\item{\code{dat}}{A \code{tbl} object, representing the linked peak cluster with differential analysis comparison/direction (cmp/direction), percentage/count (pct/cnt) of overlap, a peak count within each cluster, and a count of differential expressed genes/transcripts for such a comparison and direction. Generated by limmaConst function using limma differential analysis file and gene2peak object.}
}}

