% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHIPRNAgeneric.R, R/CHIPRNAmethods.R
\docType{methods}
\name{plotBar}
\alias{plotBar}
\alias{plotBar,tpm4plot-method}
\title{plot stacked barplots of counts of genes expressed less than 5 TPM vs more than 5 TPM.}
\usage{
plotBar(tpm.obj, pdffout)

\S4method{plotBar}{tpm4plot}(tpm.obj, pdffout)
}
\arguments{
\item{tpm.obj}{An \code{tpm4plot} object.}

\item{pdffout}{An output pdf file.}
}
\description{
plot stacked barplots of counts of genes expressed less than 5 TPM vs more than 5 TPM.
}
