% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHIPRNAgeneric.R, R/CHIPRNAplot.R
\docType{methods}
\name{plotBox}
\alias{plotBox}
\alias{plotBox,tpm4plot-method}
\title{plot boxplots of gene expressions positioned at gene expression samples separated by the identity of peak cluster membership.}
\usage{
plotBox(tpm.obj, pdffout, option)

\S4method{plotBox}{tpm4plot}(tpm.obj, pdffout, option)
}
\arguments{
\item{tpm.obj}{An \code{tpm4plot} object.}

\item{pdffout}{A pdf output file.}
}
\description{
plot boxplots of gene expressions positioned at gene expression samples separated by the identity of peak cluster membership.
}
