% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHIPRNAgeneric.R, R/CHIPRNAplot.R
\docType{methods}
\name{plotHeat}
\alias{plotHeat}
\alias{plotHeat,tpm4plot-method}
\title{plot heatmap of genes cloest to peaks}
\usage{
plotHeat(tpm.obj, pdffout, nms)

\S4method{plotHeat}{tpm4plot}(tpm.obj, pdffout, nms)
}
\arguments{
\item{tpm.obj}{A code{tpm4plot} object}

\item{pdffout}{A pdf output file}

\item{nms}{A character vector of length one, specifying directory and prefix of plotting}
}
\value{
cluster membership for peaks correspond to row scaled TPM
}
\description{
plot heatmap of genes cloest to peaks
}
