% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHIPRNAgeneric.R, R/CHIPRNAplot.R
\docType{methods}
\name{plotHeatComb}
\alias{plotHeatComb}
\alias{plotHeatComb,chip,tpm4plot-method}
\title{Plot binarized heatmap of \code{chiploj} object and gene expression heatmap in one.}
\usage{
plotHeatComb(chip.obj, tpm.obj, pdffout, option)

\S4method{plotHeatComb}{chip,tpm4plot}(chip.obj, tpm.obj, pdffout, option)
}
\arguments{
\item{chip.obj}{A code{chip} object}

\item{tpm.obj}{A code{tpm4plot} object.}

\item{pdffout}{A pdf output file.}

\item{option}{A numeric value specifying the type of tpm values for plotting. 1: row scaled TPM values; 2: all scaled TPM values; 3: log2 TPM values.}
}
\description{
Plot binarized heatmap of \code{chiploj} object and gene expression heatmap in one.
}
