% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHIPRNAgeneric.R, R/CHIPRNAplot.R
\docType{methods}
\name{plotLimma}
\alias{plotLimma}
\alias{plotLimma,limma-method}
\title{Plot peak intersected percentages of differentially expressed genes as a bar plot.}
\usage{
plotLimma(limma.obj, pdffout)

\S4method{plotLimma}{limma}(limma.obj, pdffout)
}
\arguments{
\item{limma.obj}{A code{limma} object}

\item{pdffout}{A pdf output file.}
}
\description{
Plot peak intersected percentages of differentially expressed genes as a bar plot.
}
