% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHIPRNAgeneric.R, R/CHIPRNAmethods.R
\docType{methods}
\name{updateChip}
\alias{updateChip}
\alias{updateChip,chip,gene2peak-method}
\alias{updateChip,chip,tpm4plot-method}
\title{Update the \code{bed} slot of the \code{chip} object based upon a \code{gene2peak} object or a \code{tpm4plot} object.}
\usage{
updateChip(chip.obj, other.obj)

\S4method{updateChip}{chip,gene2peak}(chip.obj, other.obj)

\S4method{updateChip}{chip,tpm4plot}(chip.obj, other.obj)
}
\arguments{
\item{chip.obj}{A \code{chip} object.}

\item{other.obj}{A \code{gene2peak} object or a \code{tpm4plot} object.}
}
\value{
A code{chip} object.
}
\description{
Update the \code{bed} slot of the \code{chip} object based upon a \code{bed} slot of a \code{gene2peak} object or a \code{var.idx} slot of a \code{tpm4plot} object.
}
