% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHIPRNAgeneric.R, R/CHIPRNAmethods.R
\docType{methods}
\name{updateChipLOJ}
\alias{updateChipLOJ}
\alias{updateChipLOJ,chiploj,gene2peak-method}
\alias{updateChipLOJ,chiploj,tpm4plot-method}
\title{Update the \code{bed} slot of the \code{chiploj} object based upon a \code{gene2peak} object or a \code{tpm4plot} object.}
\usage{
updateChipLOJ(chiploj.obj, other.obj)

\S4method{updateChipLOJ}{chiploj,gene2peak}(chiploj.obj, other.obj)

\S4method{updateChipLOJ}{chiploj,tpm4plot}(chiploj.obj, other.obj)
}
\arguments{
\item{chiploj.obj}{A \code{chiploj} object.}

\item{other.obj}{A \code{gene2peak} object or a \code{tpm4plot} object.}
}
\value{
A code{chiploj} object.
}
\description{
Update the \code{bed} slot of the \code{chiploj} object based upon a \code{gene2peak} object or a \code{tpm4plot} object.
}
