% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHIPRNAgeneric.R, R/CHIPRNAmethods.R
\docType{methods}
\name{updateG2P}
\alias{updateG2P}
\alias{updateG2P,chip,gene2peak-method}
\title{Update the \code{bed} slot of the \code{gene2peak} object based upon a \code{chip} object.}
\usage{
updateG2P(chip.obj, gene2peak.obj)

\S4method{updateG2P}{chip,gene2peak}(chip.obj, gene2peak.obj)
}
\arguments{
\item{chip.obj}{A \code{chip} object.}

\item{gene2peak.obj}{A \code{gene2peak} object.}
}
\value{
A code{gene2peak} object.
}
\description{
Update the \code{bed} slot of the \code{gene2peak} object based upon a \code{chip} object.
}
