% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHIPRNAconst.R
\name{chipConst}
\alias{chipConst}
\title{create a \code{chip} object from args list}
\usage{
chipConst(chip, sms)
}
\arguments{
\item{chip}{bed file(s)}

\item{sms}{sample name(s) for corresponding bed file(s)}
}
\value{
a \code{chip} object
}
\description{
create a \code{chip} object from args list
}
