% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHIPRNAconst.R
\name{tpmConst}
\alias{tpmConst}
\title{TPM contrast}
\usage{
tpmConst(tpm, gene2peak.obj, small = 0.05, logit = TRUE)
}
\arguments{
\item{tpm}{a \code{tpm} object.}

\item{gene2peak.obj}{a \code{gene2peak} object.}

\item{small}{a numeric value for addding to TPM value}

\item{logit}{a logical value for whethether to log2 transform TPM value before proceeding to tpm.ascale and tpm.rscale}
}
\value{
A \code{tpm4plot} object
}
\description{
Do something
}
